--- src/auth/helpers.ts
@@ -1,10 +1,0 @@
-"use server";
-import { signIn as naSignIn, signOut as naSignOut } from ".";
-
-export async function signIn() {
-	await naSignIn();
-}
-
-export async function signOut() {
-	await naSignOut();
-}
