+++ src/components/providers/sessionProvider.ts
@@ -1,0 +1,3 @@
+"use client";
+import { SessionProvider } from "next-auth/react";
+export default SessionProvider;
