+++ src/server/auth/index.ts
@@ -1,6 +1,5 @@
 import NextAuth from "next-auth";
 import { cache } from "react";
-
 import { authConfig } from "./config";
 
 const { auth: uncachedAuth, handlers, signIn, signOut } = NextAuth(authConfig);
